setup of new schema with username and password (currently not encrypted). Adapted functionalities respectively

+++ client-development/src/components/buttons-payment/AddPaymentButton.tsx
@@ -55,7 +55,7 @@ export default function AddPaymentButton(props: Props) {
     const handleChangeAmount = (e: any) => {
         setState({
             ...state,
-            amount: parseInt(e.target.value)
+            amount: parseFloat(e.target.value)
         });
     }
